2nd Attempt at Minimization Bug Fix
-Incorrect registering of a WINDOW_CLOSING event with super was causing
the window to close; corrected this to WINDOW_ICONIFIED
-doMinimize() was also closing the window, removed this call.

NOTES: It seems that registering the WINDOW_CLOSED event was disposing
of the window, and if that didn't take care of it then doMinimize()
would dispose of it.  I'm not really sure what the purpose of
doMinimize() is - what it seems to be doing is saving the state of the
program and the disposing of the window - so I just removed it.  This
seems to fix the problem and doesn't have any negative ramifications as
far as I can see.