Working Hours now seems to save to disk properly
Weekly working hours entered in the "New Project" window are now saved
to disk in the .projects file by Memoranda.

-Added get and set for WorkingHours in Project classes
-Changed createProject methods to include WorkingHours
-New Project dialog now saves the entered working hours
-Method added to Util to parse WorkingHours from a string
-Additional constructors added to WorkingHours (default and from
string), as well as a toString method (needed in ProjectImpl.java
WorkingHours set method)

WARNING: Code changed in ProductPackager.java is just a placeholder so
the program will compile; should be changed to actually read in the
saved working hours!