WorkingHours are now saved with project pack
Using File > Pack project... to export a project to a .zip file and File
> Unpack project... to import a project from a .zip file now properly
exports and imports the WorkingHours data.

- Also, changed the delimiter to a space instead of a dash, as I was
concerned that the parser would interpret the dash as a negative sign.

NOTE: There was an odd error that I was getting sometimes when trying to
pack a project; it was an array index out of bounds error in
ProductPackager's PackDirectory method:
"java.lang.ArrayIndexOutOfBoundsException: 0
at net.sf.memoranda.util.ProjectPackager.PackDirectory(Unknown Source)"
Upon repeated testing, I could not duplicate the error and packing and
unpacking worked fine, so for now I am doing no further on this issue.