Unit Tests Functioning
Unit Tests have been added and are functioning.

There is a lot of code duplication going on here however (the method
bodies are all pretty much the same except for a few integers being
changed around), so perhaps some improvement could be made there.

Also, there probably is room for much more thorough testing here;
however I think the tests I made cover the most important points well
enough and it is not worth it to spend more time on this portion of
code.