Fixed issue with auto-aggregation
I believe this issue was present in master, and when I merged master's
commits (in order to fix a Travis-CI failure) I found this issue was now
present in this branch.  The issue, specifically, was that a null
pointer exception was being thrown from TaskListImpl.CreateTask() during
the execution of the TaskListTest JUnit test.  The reason why is not
entirely clear to me, however I replaced some code I thought was a bit
dodgy with code that I think is a bit clearer (which fixed the issue) -
so I don't think it's an problem that we don't know where the null
pointer issue was coming from as long as the new code works and appears
correct.  Let me know if you see a problem with the new code.